+++ package.json
@@ -7,7 +7,7 @@
     "@apollo/client": "^3.6.9",
     "@growthbook/growthbook-react": "^0.18.0",
     "@react-oauth/google": "^0.11.1",
-    "@ribon.io/shared": "^1.1.7",
+    "@ribon.io/shared": "^1.1.8",
     "@sentry/react": "^6.16.1",
     "@stripe/react-stripe-js": "^2.1.1",
     "@stripe/stripe-js": "^1.54.1",
