+++ src/config/routes/index.tsx
@@ -2,7 +2,6 @@ import React, { Suspense, useEffect } from "react";
 import { Switch, Route } from "react-router-dom";
 import WalletProvider from "contexts/walletContext";
 import CausesPage from "pages/donations/CausesPage";
-import ConfirmDonationPage from "pages/donations/ConfirmDonationPage";
 import DonationDonePage from "pages/donations/DonationDonePage";
 import DonationDoneCausePage from "pages/donations/DonationDoneCausePage";
 import ImpactPage from "pages/users/ImpactPage";
@@ -39,10 +38,12 @@ import ContributionCanceledPage from "pages/promoters/ContributionCanceledPage";
 import ExperimentRouteComponent from "services/growthbook/ExperimentRouteComponent";
 import CampaignPage from "pages/campaigns/CampaignPage";
 import PixInstructionsPage from "pages/promoters/CheckoutPage/Components/PixInstructionsPage";
-import InsertEmailAccountPage from "pages/donations/ConfirmDonationPage/InsertEmailAccountPage";
-import ExtraTicketPage from "pages/donations/ExtraTicketPage";
-import ReceiveExtraTicketPage from "pages/donations/ReceiveExtraTicketPage";
-import AuthPage from "pages/users/AuthPage";
+import InsertEmailAccountPage from "pages/donations/auth/InsertEmailAccountPage";
+import ExtraTicketPage from "pages/donations/auth/ExtraTicketPage";
+import ReceiveExtraTicketPage from "pages/donations/auth/ReceiveExtraTicketPage";
+import SignInByTokenPage from "pages/donations/auth/SignInByTokenPage";
+import SignInPage from "pages/donations/auth/SignInPage";
+import SignedInPage from "pages/donations/auth/SignedInPage";
 import NavigationBackHeader from "./Navigation/NavigationBackHeader";
 
 function RoutesComponent(): JSX.Element {
@@ -67,7 +68,7 @@ function RoutesComponent(): JSX.Element {
 
       <Route path="/auth">
         <Suspense fallback={<div />}>
-          <AuthPage />
+          <SignInByTokenPage />
         </Suspense>
       </Route>
 
@@ -111,10 +112,17 @@ function RoutesComponent(): JSX.Element {
         </Suspense>
       </Route>
 
-      <Route path="/confirm-donation" exact>
+      <Route path="/sign-in" exact>
         <Suspense fallback={<div />}>
           <NavigationBackHeader />
-          <ConfirmDonationPage />
+          <SignInPage />
+        </Suspense>
+      </Route>
+
+      <Route path="/signed-in" exact>
+        <Suspense fallback={<div />}>
+          <NavigationBackHeader />
+          <SignedInPage />
         </Suspense>
       </Route>
