+++ src/pages/donations/auth/ExtraTicketPage/styles.ts
