+++ src/pages/donations/auth/SignInPage/GoogleSection/index.tsx
@@ -14,7 +14,7 @@ type Props = {
 };
 function GoogleSection({ onContinue }: Props): JSX.Element {
   const { t } = useTranslation("translation", {
-    keyPrefix: "donations.confirmDonationPage.signInSection",
+    keyPrefix: "donations.auth.signInPage",
   });
 
   const { signInWithGoogle } = useAuthentication();
